chore(react): reorganize theme imports and remove unused theme index file
